+++ tooling/bundle-content-scripts.ts
@@ -45,6 +45,7 @@ buildContentScript(
     getFilesInDir(`${sourceDir}/js/common/api/shared`, /\.js$/, false),
     getFilesInDir(`${sourceDir}/js/common/api/key-server`, /\.js$/, false),
     getFilesInDir(`${sourceDir}/js/common/api/account-servers`, /\.js$/, false),
+    getFilesInDir(`${sourceDir}/js/common/api/authentication`, /\.js$/, false, true),
     getFilesInDir(`${sourceDir}/js/common/api/email-provider`, /\.js$/, false),
     getFilesInDir(`${sourceDir}/js/common/api/email-provider/gmail`, /\.js$/, false),
     getFilesInDir(`${sourceDir}/js/common/api`, /\.js$/, false),
