+++ tooling/utils/tooling-utils.ts
@@ -6,9 +6,12 @@ import * as path from 'path';
 
 import { readdirSync, statSync } from 'fs';
 
-export const getFilesInDir = (dir: string, filePattern: RegExp, recursive = true): string[] => {
+export const getFilesInDir = (dir: string, filePattern: RegExp, recursive = true, reverseFilesInDir = false): string[] => {
   const all: string[] = [];
   const filesInDir = readdirSync(dir);
+  if (reverseFilesInDir) {
+    filesInDir.reverse();
+  }
   for (const fileInDir of filesInDir) {
     const filePath = path.join(dir, fileInDir);
     if (statSync(filePath).isDirectory() && recursive) {
